pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "AI Agent Workspace with Sandboxed Code Execution"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.10.0",
    "httpx>=0.28.0",
    "workos>=4.0.0",
    "daytona-sdk>=0.10.5",
    "python-multipart>=0.0.18",
    "sse-starlette>=2.2.0",
    "websockets>=14.0",
    "openai>=1.60.0",
    "serpapi>=0.1.5",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
